+++ src/main.cpp
@@ -5,60 +5,60 @@
 
 using namespace geode::prelude;
 
-class $modify(OTPauseLayer, PauseLayer) {
-	void customSetup() {
-		this->PauseLayer::customSetup();
+class OverTherePopup : public Popup {
+protected:
+	bool init() {
+		if (!Popup::init(210.f, 120.f)) return false;
+		this->setTitle("Over There Options");
+		this->setID("over-there-popup");
 
-		// create nodes for labels
-		auto otItem1 = CCNode::create();
-		otItem1->setContentSize({50.f, 26.f});
-		auto otItem2 = CCNode::create();
-		otItem2->setContentSize({50.f, 40.f});
+		auto menu = CCMenu::create();
+		menu->setPosition({0.f, 0.f});
+		menu->setID("over-there-menu");
+		m_mainLayer->addChild(menu);
 
 		// create toggle box 1 
 		auto offSpr = CCSprite::createWithSpriteFrameName("GJ_checkOff_001.png");
 		auto onSpr = CCSprite::createWithSpriteFrameName("GJ_checkOn_001.png");
-		auto otToggle1 = CCMenuItemToggler::create(offSpr, onSpr, this, menu_selector(OTPauseLayer::toggleOT));
+		auto otToggle1 = CCMenuItemToggler::create(offSpr, onSpr, this, menu_selector(OverTherePopup::toggleOT));
 		otToggle1->setAnchorPoint({0.5f, 0.5f});
-		otToggle1->setID("over-there-toggle");
+		otToggle1->setPosition({140.f, 68.f});
+		otToggle1->setID("over-there-toggle1");
+		menu->addChild(otToggle1);
 
 		// create toggle box 2
-		auto otToggle2 = CCMenuItemToggler::create(offSpr, onSpr, this, menu_selector(OTPauseLayer::togglePracticeOT));
+		auto otToggle2 = CCMenuItemToggler::create(offSpr, onSpr, this, menu_selector(OverTherePopup::togglePracticeOT));
 		otToggle2->setAnchorPoint({0.5f, 0.5f});
+		otToggle2->setPosition({70.f, 68.f});
 		otToggle2->setID("over-there-toggle2");
+		menu->addChild(otToggle2);
 
 		// apply setting values to toggle boxes
 		bool togglePointers = Mod::get()->getSettingValue<bool>("toggle-pointers");
 		otToggle1->toggle(togglePointers ? 1 : 0);
 		bool practicePointers = Mod::get()->getSettingValue<bool>("practice-pointers");
 		otToggle2->toggle(practicePointers ? 1 : 0);
+
 		
 		// create toggle box label
 		auto otLabel = CCLabelBMFont::create("Toggle\nPointers", "bigFont.fnt", 0.f, kCCTextAlignmentCenter);
-		otLabel->setPosition({25.f, 13.f});
+		otLabel->setPosition({140.f, 35.f});
 		otLabel->setScale(0.3f);
 		otLabel->setID("over-there-label");
+		menu->addChild(otLabel);
 
 		// create practice mode label
 		auto otPracticeLabel = CCLabelBMFont::create("Pointers\nOnly In\nPractice", "bigFont.fnt", 0.f, kCCTextAlignmentCenter);
-		otPracticeLabel->setPosition({25.f, 20.f});
+		otPracticeLabel->setPosition({70.f, 30.f});
 		otPracticeLabel->setScale(0.3f);
 		otPracticeLabel->setID("over-there-practice-label");
+		menu->addChild(otPracticeLabel);
 
-		// add labels to their nodes
-		otItem1->addChild(otLabel);
-		otItem2->addChild(otPracticeLabel);
-		
-		if (auto menu = this->getChildByID("right-button-menu")) {
-			menu->addChild(otToggle2);
-			menu->addChild(otItem2);
-			menu->addChild(otToggle1);
-			menu->addChild(otItem1);
-			menu->updateLayout();
-		} // obligatory comment
-	} // epic comment
+		return true;
+	} // init
 
 	// callback functions
+
 	void toggleOT(CCObject* sender) {
 		auto toggle = static_cast<CCMenuItemToggler*>(sender);
 		bool state = !toggle->isToggled();
@@ -66,13 +66,50 @@ class $modify(OTPauseLayer, PauseLayer) {
 		return;
 	} // toggle function
 
+
 	void togglePracticeOT(CCObject* sender) {
 		auto toggle = static_cast<CCMenuItemToggler*>(sender);
 		bool state = !toggle->isToggled();
 		Mod::get()->setSettingValue<bool>("practice-pointers", state);
 		return;
 	} // practice function
-}; // UI creation is butt-ugly
+public:
+	static OverTherePopup* create() {
+		auto ret = new OverTherePopup();
+		if (ret && ret->init()) {
+			ret->autorelease();
+			return ret;
+		}
+		delete ret;
+		return nullptr;
+	} // create
+};
+
+
+class $modify(OTPauseLayer, PauseLayer) {
+	void customSetup() {
+		this->PauseLayer::customSetup();
+
+		// create button
+		auto spr = CCSprite::create("OT_logoBtn_001.png"_spr);
+		spr->setScale(0.65f);
+		auto btn = CCMenuItemSpriteExtra::create(spr, this, menu_selector(OTPauseLayer::onLogoBtn));
+		btn->setAnchorPoint({0.5f, 0.5f});
+		btn->setID("over-there-button");
+		
+		// add button
+		if (auto menu = this->getChildByID("right-button-menu")) {
+			menu->addChild(btn);
+			menu->updateLayout();
+		}
+	} // custom setup
+
+	// callback functions
+
+	void onLogoBtn(CCObject *) {
+		OverTherePopup::create()->show();
+	} // comment
+};
 
 
 class $modify(OTPlayLayer, PlayLayer) {
@@ -151,12 +188,11 @@ class $modify(OTPlayLayer, PlayLayer) {
 		auto arrow = static_cast<CCSprite*>(entity->getChildByID("arrow"));
 		auto cross = static_cast<CCSprite*>(entity->getChildByID("cross"));
 		float cameraZoom = GJBaseGameLayer::get()->m_gameState.m_cameraZoom;
-		CCPoint cameraOffset = GJBaseGameLayer::get()->m_gameState.m_cameraOffset;
 		auto winSize = CCDirector::sharedDirector()->getWinSize();
 		auto lootPos = target->getLastPosition();
 		CCPoint cam = GJBaseGameLayer::get()->m_gameState.m_cameraPosition;
 		float cameraFlip = GJBaseGameLayer::get()->m_cameraFlip;
-		CCPoint screenPos = ccp((lootPos.x - cam.x) * cameraZoom * cameraFlip + cameraOffset.x, (lootPos.y - cam.y) * cameraZoom + cameraOffset.y);
+		CCPoint screenPos = ccp((lootPos.x - cam.x) * cameraZoom * cameraFlip, (lootPos.y - cam.y) * cameraZoom);
 		float distanceSq;
 		float onScreenMargin = 20.f;
 		CCPoint distance = ccpSub(cam, lootPos);
@@ -231,9 +267,8 @@ class $modify(OTPlayLayer, PlayLayer) {
 		auto lootPos = target->getLastPosition();
 		CCPoint cam = GJBaseGameLayer::get()->m_gameState.m_cameraPosition;
 		float cameraZoom = GJBaseGameLayer::get()->m_gameState.m_cameraZoom;
-		CCPoint cameraOffset = GJBaseGameLayer::get()->m_gameState.m_cameraOffset;
 		float cameraFlip = GJBaseGameLayer::get()->m_cameraFlip;
-		CCPoint screenPos = ccp((lootPos.x - cam.x) * cameraZoom * cameraFlip + cameraOffset.x, (lootPos.y - cam.y) * cameraZoom + cameraOffset.y);
+		CCPoint screenPos = ccp((lootPos.x - cam.x) * cameraZoom * cameraFlip, (lootPos.y - cam.y) * cameraZoom);
 		auto winSize = CCDirector::sharedDirector()->getWinSize();
 
 		if (cameraFlip < 0) {